feat: 实现设备管理服务核心功能与架构优化
refactor(auth): 重构用户认证与角色服务接口
fix(device): 修复设备连接与属性服务问题
feat(datacenter): 新增设备数据同步与存储服务
docs: 更新数据采集中心文档与配置说明
style: 统一代码风格与命名规范
test: 添加缺失的测试用例
chore: 更新依赖与配置文件

+++ device/data/kv_event.go
@@ -5,7 +5,7 @@ import "gorm.io/gorm"
 type KvEvent struct {
 	//ID         int64  `json:"id" gorm:"column:id"`
 	gorm.Model
-	ProductID  int    `json:"product_id" gorm:"column:product_id"`
+	ProductID  int64  `json:"product_id" gorm:"column:product_id"` // 修改为int64以保持一致
 	ProductKey string `json:"product_key" gorm:"column:product_key"` // 产品标识
 	Name       string `json:"name" gorm:"column:name"`               // 动作名称
 	Identifier string `json:"identifier" gorm:"column:identifier"`   // 动作标识符
